feat: refactor cart components and implement new cart view
- Updated the cart page to use the new `CartView` component for improved structure and functionality.
- Refactored import paths for `useCartActions` to align with the new directory structure.
- Introduced `CartItemCard` and `QuantityButtons` components to enhance the cart item display and interaction.
- Added a new `useCartActions` hook for managing cart actions, including adding, removing, and updating item quantities.
