+++ components/shared/sections/Collection.tsx
@@ -2,7 +2,7 @@
 
 import ProductCard from "@/components/shared/ProductCard";
 import Title from "@/components/shared/Title";
-import { useCartActions } from "@/hooks/useCartActions";
+import { useCartActions } from "@/features/cart/hooks/useCartActions";
 import { useFavoriteActions } from "@/hooks/useFavoriteActions";
 import { useProduct } from "@/hooks/useProduct";
 import { useFavoriteStore } from "@/stores/FavoriteStore";
